requirements.txt
transformers
datasets
sentence-transformers
evaluate
jiwer
soundfile
accelerate>=0.26.0
bert-score
torch
evaluate
pyyaml
tensorboard